+++ nucliadb_client/nucliadb_client/knowledgebox.py
@@ -18,9 +18,19 @@
 # along with this program. If not, see <http://www.gnu.org/licenses/>.
 
 import base64
+from enum import Enum
 from typing import TYPE_CHECKING, AsyncIterator, List, Optional
 
-from nucliadb_protos.writer_pb2 import BrokerMessage, ExportRequest
+from nucliadb_protos.writer_pb2 import (
+    BrokerMessage,
+    ExportRequest,
+    GetEntitiesRequest,
+    GetEntitiesResponse,
+    GetLabelsRequest,
+    GetLabelsResponse,
+    SetEntitiesRequest,
+    SetLabelsRequest,
+)
 
 from nucliadb_models.resource import KnowledgeBoxObj, ResourceList
 from nucliadb_models.writer import CreateResourcePayload, ResourceCreated
@@ -28,11 +38,18 @@ from nucliadb_models.writer import CreateResourcePayload, ResourceCreated
 if TYPE_CHECKING:
     from nucliadb_client.client import NucliaDBClient
 
+import aiofiles
 import httpx
 
 from nucliadb_client.resource import Resource
 
 KB_PREFIX = "kb"
+
+
+class CODEX(str, Enum):
+    RES = "RES:"
+    LAB = "LAB:"
+    ENT = "ENT:"
 
 
 class KnowledgeBox:
@@ -83,19 +100,71 @@ class KnowledgeBox:
         resp = self.http_manager_v1.delete("")
         return resp.status_code == 200
 
-    def parse_bm(self, payload: bytes) -> Resource:
-        pb = BrokerMessage()
-        pb.ParseFromString(payload)
-        res = Resource(rid=pb.uuid, kb=self)
-        res._bm = pb
-        return res
+    async def import_export(self, line: str):
+        type_line = line[:4]
+        payload = base64.b64decode(line[4:])
+        if type_line == CODEX.RES:
+            pb_bm = BrokerMessage()
+            pb_bm.ParseFromString(payload)
+            res = Resource(rid=pb_bm.uuid, kb=self)
+            res._bm = pb_bm
+            await res.commit(processor=False)
+        elif type_line == CODEX.ENT:
+            pb_er = GetEntitiesResponse()
+            pb_er.ParseFromString(payload)
+            for group, entities in pb_er.groups.items():
+                ser_pb = SetEntitiesRequest()
+                ser_pb.kb.uuid = self.kbid
+                ser_pb.group = group
+                ser_pb.entities.CopyFrom(entities)
+                await self.client.writer_stub_async.SetEntities(ser_pb)  # type:  ignore
 
-    async def export(self) -> AsyncIterator[str]:
+        elif type_line == CODEX.LAB:
+            pb_lr = GetLabelsResponse()
+            pb_lr.ParseFromString(payload)
+            for labelset, labelset_obj in pb_lr.labels.labelset.items():
+                slr_pb = SetLabelsRequest()
+                slr_pb.kb.uuid = self.kbid
+                slr_pb.id = labelset
+                slr_pb.labelset.CopyFrom(labelset_obj)
+                await self.client.writer_stub_async.SetLabels(slr_pb)  # type:  ignore
+
+    async def resources(self) -> AsyncIterator[BrokerMessage]:
         assert self.client.writer_stub_async
         req = ExportRequest()
         req.kbid = self.kbid
         async for bm in self.client.writer_stub_async.Export(req):  # type: ignore
-            yield base64.b64encode(bm.SerializeToString()).decode()
+            yield bm
+
+    async def entities(self) -> GetEntitiesResponse:
+        assert self.client.writer_stub_async
+        req = GetEntitiesRequest()
+        req.kb.uuid = self.kbid
+        entities_response: GetEntitiesResponse = await self.client.writer_stub_async.GetEntities(req)  # type: ignore
+        return entities_response
+
+    async def labels(self) -> GetLabelsResponse:
+        assert self.client.writer_stub_async
+        req = GetLabelsRequest()
+        req.kb.uuid = self.kbid
+        label_response: GetLabelsResponse = (
+            await self.client.writer_stub_async.GetLabels(req)  # type: ignore
+        )
+        return label_response
 
     def init_async_grpc(self):
         self.client.init_async_grpc()
+
+    async def generator(self) -> AsyncIterator[str]:
+        self.init_async_grpc()
+        async for bm in self.resources():
+            yield CODEX.RES + base64.b64encode(bm.SerializeToString()).decode() + "\n"
+        entities = await self.entities()
+        yield CODEX.ENT + base64.b64encode(entities.SerializeToString()).decode() + "\n"
+        labels = await self.labels()
+        yield CODEX.LAB + base64.b64encode(labels.SerializeToString()).decode() + "\n"
+
+    async def export(self, dump: str):
+        async with aiofiles.open(dump, "w+") as dump_file:
+            async for line in self.generator():
+                await dump_file.write(line)